+++ common/src/main/java/net/conczin/mca/ducks/client/PlayerRendererMCA.java
@@ -1,9 +1,0 @@
-package net.conczin.mca.ducks.client;
-
-import com.mojang.blaze3d.vertex.PoseStack;
-import net.minecraft.client.player.AbstractClientPlayer;
-import net.minecraft.client.renderer.SubmitNodeCollector;
-
-public interface PlayerRendererMCA {
-    boolean mca$renderHand(AbstractClientPlayer player, PoseStack poseStack, SubmitNodeCollector submitNodeCollector, int lightCoords, boolean rightArm, boolean hasSleeve);
-}