+++ common/src/main/java/net/conczin/mca/network/ClientHandlerImpl.java
@@ -240,6 +240,7 @@ public class ClientHandlerImpl implements ClientHandler {
     @Override
     public void handleConfigResponse(ConfigResponse message) {
         Config.setServerConfig(message.getConfig());
+        MCAClient.refreshPlayerDataDependentDimensions();
     }
 
     @Override